Implementation of Module#3 & Module#5 + Some Minor Changes in Schemas
Made a new Model for Wishlist and updates User Model with an additional attribute called role which can be either admin or user.

Add methods to register and login method

Some Minor changes in userController.js methods

Some Minor changes in userRoutes.js

Added and extra middleware called checkRole - it decodes JWT token and checks if the role of the request is admin or user.